cloud/api/requirements.txt
# Cloud API Dependencies
fastapi>=0.109.0
uvicorn[standard]>=0.25.0
python-multipart>=0.0.6
aiohttp>=3.9.1
pydantic>=2.5.3
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
pyyaml>=6.0.1

# Database (future)
sqlalchemy>=2.0.25
alembic>=1.13.1
asyncpg>=0.29.0
redis>=5.0.1

# Monitoring
psutil>=5.9.7
